first impolementation of useSideBarOrderedReports decompoostion
