+++ src/hooks/useSidebarOrderedReports.tsx
@@ -1,44 +1,24 @@
 import {deepEqual} from 'fast-equals';
-import React, {createContext, useCallback, useContext, useEffect, useMemo, useRef, useState} from 'react';
-import type {OnyxEntry} from 'react-native-onyx';
+import React, {createContext, useContext, useEffect, useRef} from 'react';
 import Log from '@libs/Log';
-import {getTransactionThreadReportID} from '@libs/MergeTransactionUtils';
-import {getPolicyEmployeeListByIdWithoutCurrentUser} from '@libs/PolicyUtils';
-import {isOneTransactionReport} from '@libs/ReportUtils';
-import SidebarUtils from '@libs/SidebarUtils';
-import CONST from '@src/CONST';
-import ONYXKEYS from '@src/ONYXKEYS';
-import type * as OnyxTypes from '@src/types/onyx';
-import {useCurrentReportIDState} from './useCurrentReportID';
-import useCurrentUserPersonalDetails from './useCurrentUserPersonalDetails';
-import useLocalize from './useLocalize';
-import useMappedPolicies from './useMappedPolicies';
-import useOnyx from './useOnyx';
 import usePrevious from './usePrevious';
 import useReportAttributes from './useReportAttributes';
-import useResponsiveLayout from './useResponsiveLayout';
+import type {ReportsToDisplayInLHN} from './useReportsToDisplayInLHN';
+import useReportsToDisplayInLHN from './useReportsToDisplayInLHN';
+import useSidebarData from './useSidebarData';
+import type {SidebarOrderedReportsStateContextValue} from './useSortedLHNReports';
+import useSortedLHNReports from './useSortedLHNReports';
 
 const componentsUsingHook = new Map<string, {renderDuration: number}>();
-
-type PartialPolicyForSidebar = Pick<OnyxTypes.Policy, 'type' | 'name' | 'avatarURL' | 'employeeList'>;
 
 type SidebarOrderedReportsContextProviderProps = {
     children: React.ReactNode;
     currentReportIDForTests?: string;
 };
 
-type SidebarOrderedReportsStateContextValue = {
-    orderedReports: OnyxTypes.Report[];
-    orderedReportIDs: string[];
-    currentReportID: string | undefined;
-    policyMemberAccountIDs: number[];
-};
-
 type SidebarOrderedReportsActionsContextValue = {
     clearLHNCache: () => void;
 };
-
-type ReportsToDisplayInLHN = Record<string, OnyxTypes.Report & {hasErrorsOtherThanFailedReceipt?: boolean; requiresAttention?: boolean}>;
 
 const SidebarOrderedReportsStateContext = createContext<SidebarOrderedReportsStateContextValue>({
     orderedReports: [],
@@ -50,14 +30,6 @@ const SidebarOrderedReportsStateContext = createContext<SidebarOrderedReportsSta
 const SidebarOrderedReportsActionsContext = createContext<SidebarOrderedReportsActionsContextValue>({
     clearLHNCache: () => {},
 });
-
-const policyMapper = (policy: OnyxEntry<OnyxTypes.Policy>): PartialPolicyForSidebar =>
-    (policy && {
-        type: policy.type,
-        name: policy.name,
-        avatarURL: policy.avatarURL,
-        employeeList: policy.employeeList,
-    }) as PartialPolicyForSidebar;
 
 function SidebarOrderedReportsContextProvider({
     children,
@@ -71,283 +43,55 @@ function SidebarOrderedReportsContextProvider({
      */
     currentReportIDForTests,
 }: SidebarOrderedReportsContextProviderProps) {
-    const {localeCompare} = useLocalize();
-    const [priorityMode = CONST.PRIORITY_MODE.DEFAULT] = useOnyx(ONYXKEYS.NVP_PRIORITY_MODE);
-    const [chatReports, {sourceValue: reportUpdates}] = useOnyx(ONYXKEYS.COLLECTION.REPORT);
-    const [policies, {sourceValue: policiesUpdates}] = useMappedPolicies(policyMapper);
-    const [transactions, {sourceValue: transactionsUpdates}] = useOnyx(ONYXKEYS.COLLECTION.TRANSACTION);
-    const [transactionViolations, {sourceValue: transactionViolationsUpdates}] = useOnyx(ONYXKEYS.COLLECTION.TRANSACTION_VIOLATIONS);
-    const [reportNameValuePairs, {sourceValue: reportNameValuePairsUpdates}] = useOnyx(ONYXKEYS.COLLECTION.REPORT_NAME_VALUE_PAIRS);
-    const [reportsDrafts, {sourceValue: reportsDraftsUpdates}] = useOnyx(ONYXKEYS.COLLECTION.REPORT_DRAFT_COMMENT);
-    const [betas] = useOnyx(ONYXKEYS.BETAS);
-    const [conciergeReportID] = useOnyx(ONYXKEYS.CONCIERGE_REPORT_ID);
+    const data = useSidebarData(currentReportIDForTests);
     const reportAttributes = useReportAttributes();
-    const [currentReportsToDisplay, setCurrentReportsToDisplay] = useState<ReportsToDisplayInLHN>({});
-    const {shouldUseNarrowLayout} = useResponsiveLayout();
-    const {accountID} = useCurrentUserPersonalDetails();
-    const {currentReportID: currentReportIDValue} = useCurrentReportIDState();
-    const derivedCurrentReportID = currentReportIDForTests ?? currentReportIDValue;
-    const prevDerivedCurrentReportID = usePrevious(derivedCurrentReportID);
-
-    // we need to force reportsToDisplayInLHN to re-compute when we clear currentReportsToDisplay, but the way it currently works relies on not having currentReportsToDisplay as a memo dependency, so we just need something we can change to trigger it
-    // I don't like it either, but clearing the cache is only a hack for the debug modal and I will endeavor to make it better as I work to improve the cache correctness of the LHN more broadly
-    const [clearCacheDummyCounter, setClearCacheDummyCounter] = useState(0);
-
-    const policyMemberAccountIDs = useMemo(() => getPolicyEmployeeListByIdWithoutCurrentUser(policies, undefined, accountID), [policies, accountID]);
-    const prevBetas = usePrevious(betas);
-    const prevPriorityMode = usePrevious(priorityMode);
-
-    const perfRef = useRef<{hookDuration: number}>({
-        hookDuration: 0,
-    });
-    const hookStartTime = useRef<number>(performance.now());
-
-    /**
-     * Find the reports that need to be updated in the LHN
-     */
-    const getUpdatedReports = useCallback(() => {
-        const reportsToUpdate = new Set<string>();
-
-        if (betas !== prevBetas || priorityMode !== prevPriorityMode) {
-            for (const key of Object.keys(chatReports ?? {})) {
-                reportsToUpdate.add(key);
-            }
-        }
-        if (reportUpdates) {
-            for (const key of Object.keys(reportUpdates ?? {})) {
-                reportsToUpdate.add(key);
-            }
-        }
-        if (reportNameValuePairsUpdates) {
-            for (const key of Object.keys(reportNameValuePairsUpdates ?? {}).map((reportKey) => reportKey.replace(ONYXKEYS.COLLECTION.REPORT_NAME_VALUE_PAIRS, ONYXKEYS.COLLECTION.REPORT))) {
-                reportsToUpdate.add(key);
-            }
-        }
-        if (transactionsUpdates) {
-            // We need to select the report linked to a transaction, to properly recalculate getReceiptUploadErrorReason, which is the expense report if it is isOneTransactionReport
-            // or the transaction thread report if it is otherwise.
-            for (const key of Object.values(transactionsUpdates ?? {}).map((transaction) =>
-                transaction?.reportID && isOneTransactionReport(chatReports?.[`${ONYXKEYS.COLLECTION.REPORT}${transaction.reportID}`])
-                    ? `${ONYXKEYS.COLLECTION.REPORT}${transaction?.reportID}`
-                    : `${ONYXKEYS.COLLECTION.REPORT}${getTransactionThreadReportID(transaction)}`,
-            )) {
-                reportsToUpdate.add(key);
-            }
-        }
-        if (transactionViolationsUpdates) {
-            for (const key of Object.keys(transactionViolationsUpdates ?? {})
-                .map((violationKey) => violationKey.replace(ONYXKEYS.COLLECTION.TRANSACTION_VIOLATIONS, ONYXKEYS.COLLECTION.TRANSACTION))
-                .map((transactionKey) => `${ONYXKEYS.COLLECTION.REPORT}${transactions?.[transactionKey]?.reportID}`)) {
-                reportsToUpdate.add(key);
-            }
-        }
-        if (reportsDraftsUpdates) {
-            for (const key of Object.keys(reportsDraftsUpdates).map((draftKey) => draftKey.replace(ONYXKEYS.COLLECTION.REPORT_DRAFT_COMMENT, ONYXKEYS.COLLECTION.REPORT))) {
-                reportsToUpdate.add(key);
-            }
-        }
-        if (policiesUpdates) {
-            const updatedPolicies = new Set(Object.keys(policiesUpdates).map((policyKey) => policyKey.replace(ONYXKEYS.COLLECTION.POLICY, '')));
-            for (const key of Object.entries(chatReports ?? {})
-                .filter(([, value]) => {
-                    if (!value?.policyID) {
-                        return;
-                    }
-
-                    return updatedPolicies.has(value.policyID);
-                })
-                .map(([reportKey]) => reportKey)) {
-                reportsToUpdate.add(key);
-            }
-        }
-
-        // Make sure the previous and current reports are always included in the updates when we switch reports.
-        if (prevDerivedCurrentReportID !== derivedCurrentReportID) {
-            reportsToUpdate.add(`${ONYXKEYS.COLLECTION.REPORT}${prevDerivedCurrentReportID}`);
-            reportsToUpdate.add(`${ONYXKEYS.COLLECTION.REPORT}${derivedCurrentReportID}`);
-        }
-
-        return Array.from(reportsToUpdate);
-    }, [
-        reportUpdates,
-        reportNameValuePairsUpdates,
-        transactionsUpdates,
-        transactionViolationsUpdates,
-        reportsDraftsUpdates,
-        policiesUpdates,
-        chatReports,
-        transactions,
-        betas,
-        priorityMode,
-        prevBetas,
-        prevPriorityMode,
-        prevDerivedCurrentReportID,
-        derivedCurrentReportID,
-    ]);
-
-    const reportsToDisplayInLHN = useMemo(() => {
-        const updatedReports = getUpdatedReports();
-        const hasCachedReports = Object.keys(currentReportsToDisplay).length > 0;
-
-        // When reportAttributes changes (e.g. on startup hydration) but no report-specific keys were
-        // updated, getUpdatedReports() returns []. Rather than falling through to a full scan of all
-        // reports, recheck only the already-displayed reports with the new reportAttributes.
-        const effectiveUpdatedReports = updatedReports.length === 0 && hasCachedReports ? Object.keys(currentReportsToDisplay) : updatedReports;
-        const shouldDoIncrementalUpdate = effectiveUpdatedReports.length > 0 && hasCachedReports;
-        let reportsToDisplay = {};
-        if (shouldDoIncrementalUpdate) {
-            reportsToDisplay = SidebarUtils.updateReportsToDisplayInLHN({
-                displayedReports: currentReportsToDisplay,
-                reports: chatReports,
-                updatedReportsKeys: effectiveUpdatedReports,
-                currentReportId: derivedCurrentReportID,
-                isInFocusMode: priorityMode === CONST.PRIORITY_MODE.GSD,
-                betas,
-                transactionViolations,
-                reportNameValuePairs,
-                reportAttributes,
-                draftComments: reportsDrafts,
-                transactions,
-            });
-        } else {
-            Log.info('[useSidebarOrderedReports] building reportsToDisplay from scratch');
-            reportsToDisplay = SidebarUtils.getReportsToDisplayInLHN(
-                derivedCurrentReportID,
-                chatReports,
-                betas,
-                priorityMode,
-                reportsDrafts,
-                transactionViolations,
-                transactions,
-                reportNameValuePairs,
-                reportAttributes,
-            );
-        }
-
-        return reportsToDisplay;
-        // Rule disabled intentionally — triggering a re-render on currentReportsToDisplay would cause an infinite loop
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [getUpdatedReports, chatReports, derivedCurrentReportID, priorityMode, betas, transactionViolations, reportNameValuePairs, reportAttributes, reportsDrafts, clearCacheDummyCounter]);
-
-    // Derive a stable boolean map indicating which reports have drafts.
-    const hasDraftByReportIDRef = useRef<Record<string, boolean>>({});
-    const hasDraftByReportID = useMemo(() => {
-        const result: Record<string, boolean> = {};
-        if (reportsDrafts) {
-            for (const [key, value] of Object.entries(reportsDrafts)) {
-                if (value) {
-                    result[key.replace(ONYXKEYS.COLLECTION.REPORT_DRAFT_COMMENT, '')] = true;
-                }
-            }
-        }
-        const prev = hasDraftByReportIDRef.current;
-        const prevKeys = Object.keys(prev);
-        const newKeys = Object.keys(result);
-        if (prevKeys.length === newKeys.length && newKeys.every((k) => k in prev)) {
-            return prev;
-        }
-        hasDraftByReportIDRef.current = result;
-        return result;
-    }, [reportsDrafts]);
-
-    useEffect(() => {
-        setCurrentReportsToDisplay(reportsToDisplayInLHN);
-    }, [reportsToDisplayInLHN]);
-
-    const getOrderedReportIDs = useCallback(
-        () => SidebarUtils.sortReportsToDisplayInLHN(reportsToDisplayInLHN, priorityMode, localeCompare, hasDraftByReportID, reportNameValuePairs, conciergeReportID),
-        // Rule disabled intentionally - reports should be sorted only when the reportsToDisplayInLHN changes
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        [reportsToDisplayInLHN, localeCompare, hasDraftByReportID, conciergeReportID],
+    const {reportsToDisplayInLHN, clearLHNCache} = useReportsToDisplayInLHN(data, reportAttributes);
+    const stateValue = useSortedLHNReports(
+        reportsToDisplayInLHN,
+        data.priorityMode,
+        data.reportsDrafts,
+        data.reportNameValuePairs,
+        data.conciergeReportID,
+        data.derivedCurrentReportID,
+        data.shouldUseNarrowLayout,
+        data.chatReports,
+        data.policies,
+        data.accountID,
     );
 
-    const orderedReportIDs = useMemo(() => getOrderedReportIDs(), [getOrderedReportIDs]);
-
-    // Get the actual reports based on the ordered IDs
-    const getOrderedReports = useCallback(
-        (reportIDs: string[]): OnyxTypes.Report[] => {
-            if (!chatReports) {
-                return [];
-            }
-            return reportIDs.map((reportID) => chatReports[`${ONYXKEYS.COLLECTION.REPORT}${reportID}`]).filter(Boolean) as OnyxTypes.Report[];
-        },
-        [chatReports],
-    );
-
-    const orderedReports = useMemo(() => getOrderedReports(orderedReportIDs), [getOrderedReports, orderedReportIDs]);
-
-    const clearLHNCache = useCallback(() => {
-        Log.info('[useSidebarOrderedReports] Clearing sidebar cache manually via debug modal');
-        setCurrentReportsToDisplay({});
-        setClearCacheDummyCounter((current) => current + 1);
-    }, []);
-
-    const stateValue: SidebarOrderedReportsStateContextValue = useMemo(() => {
-        // We need to make sure the current report is in the list of reports, but we do not want
-        // to have to re-generate the list every time the currentReportID changes. To do that
-        // we first generate the list as if there was no current report, then we check if
-        // the current report is missing from the list, which should very rarely happen. In this
-        // case we re-generate the list a 2nd time with the current report included.
-
-        // We also execute the following logic if `shouldUseNarrowLayout` is false because this is
-        // requirement for web. Consider a case, where we have report with expenses and we click on
-        // any expense, a new LHN item is added in the list and is visible on web. But on mobile, we
-        // just navigate to the screen with expense details, so there seems no point to execute this logic on mobile.
-        if (
-            (!shouldUseNarrowLayout || orderedReportIDs.length === 0) &&
-            derivedCurrentReportID &&
-            derivedCurrentReportID !== '-1' &&
-            orderedReportIDs.indexOf(derivedCurrentReportID) === -1
-        ) {
-            const updatedReportIDs = getOrderedReportIDs();
-            const updatedReports = getOrderedReports(updatedReportIDs);
-            return {
-                orderedReports: updatedReports,
-                orderedReportIDs: updatedReportIDs,
-                currentReportID: derivedCurrentReportID,
-                policyMemberAccountIDs,
-            };
-        }
-
-        return {
-            orderedReports,
-            orderedReportIDs,
-            currentReportID: derivedCurrentReportID,
-            policyMemberAccountIDs,
-        };
-    }, [getOrderedReportIDs, orderedReportIDs, derivedCurrentReportID, policyMemberAccountIDs, shouldUseNarrowLayout, getOrderedReports, orderedReports]);
-
-    const actionsValue: SidebarOrderedReportsActionsContextValue = useMemo(() => ({clearLHNCache}), [clearLHNCache]);
+    const actionsValue: SidebarOrderedReportsActionsContextValue = {clearLHNCache};
 
     const currentDeps = {
-        priorityMode,
-        chatReports,
-        policies,
-        transactions,
-        transactionViolations,
-        reportNameValuePairs,
-        betas,
+        priorityMode: data.priorityMode,
+        chatReports: data.chatReports,
+        policies: data.policies,
+        transactions: data.transactions,
+        transactionViolations: data.transactionViolations,
+        reportNameValuePairs: data.reportNameValuePairs,
+        betas: data.betas,
         reportAttributes,
-        currentReportsToDisplay,
-        shouldUseNarrowLayout,
-        accountID,
-        currentReportIDValue,
-        derivedCurrentReportID,
-        prevDerivedCurrentReportID,
-        policyMemberAccountIDs,
-        prevBetas,
-        prevPriorityMode,
+        shouldUseNarrowLayout: data.shouldUseNarrowLayout,
+        accountID: data.accountID,
+        currentReportIDValue: data.derivedCurrentReportID,
+        derivedCurrentReportID: data.derivedCurrentReportID,
+        prevDerivedCurrentReportID: data.prevDerivedCurrentReportID,
+        policyMemberAccountIDs: stateValue.policyMemberAccountIDs,
+        prevBetas: data.prevBetas,
+        prevPriorityMode: data.prevPriorityMode,
         reportsToDisplayInLHN,
-        orderedReportIDs,
-        orderedReports,
+        orderedReportIDs: stateValue.orderedReportIDs,
+        orderedReports: stateValue.orderedReports,
     };
     const prevContextValue = usePrevious(stateValue);
     const previousDeps = usePrevious(currentDeps);
     const firstRender = useRef(true);
 
+    const perfRef = useRef<{hookDuration: number}>({hookDuration: 0});
+    // eslint-disable-next-line react-hooks/purity
+    const hookStartTime = useRef<number>(performance.now());
+
     useEffect(() => {
-        const hookExecutionDuration = performance.now() - hookStartTime.current;
-        perfRef.current.hookDuration = hookExecutionDuration;
+        perfRef.current.hookDuration = performance.now() - hookStartTime.current;
     }, [stateValue]);
 
     useEffect(() => {
@@ -428,7 +172,7 @@ function useSidebarOrderedReportsPerformance(componentName?: string) {
 }
 
 export {SidebarOrderedReportsContextProvider, useSidebarOrderedReports, useSidebarOrderedReportsState, useSidebarOrderedReportsActions};
-export type {PartialPolicyForSidebar, ReportsToDisplayInLHN};
+export type {ReportsToDisplayInLHN, SidebarOrderedReportsStateContextValue};
 
 function getChangedKeys<T extends Record<string, unknown>>(deps: T, prevDeps: T) {
     const depsKeys = Object.keys(deps);
@@ -449,9 +193,9 @@ function logChangedDeps<T extends Record<string, unknown>>(msg: string, deps: T,
         performance: {
             hookDuration: `${hookExecutionDuration.toFixed(2)}ms`,
             loggingDuration: `${loggingDuration.toFixed(2)}ms`,
-            componentsUsingHook: Array.from(componentsUsingHook.entries()).map(([name, data]) => ({
+            componentsUsingHook: Array.from(componentsUsingHook.entries()).map(([name, hookData]) => ({
                 component: name,
-                renderDuration: `${data.renderDuration.toFixed(2)}ms`,
+                renderDuration: `${hookData.renderDuration.toFixed(2)}ms`,
             })),
         },
